Fix timeout type from in to float.

+++ src/elasticsearch_dsl-stubs/__init__.pyi
@@ -62,7 +62,7 @@ class connections:  # noqa: N801
         scheme: str = ...,
         use_ssl: bool = ...,
         url_prefix: Optional[str] = ...,
-        timeout: int = ...,
+        timeout: float = ...,
         # headers
         http_compress: Optional[bool] = ...,
         cloud_id: Optional[str] = ...,
